Filter out snapshot branches in the parse phase
snapshot branches should not be included in the database created by
the parse phase as it will be redundant in the export phase.
The baseline branches that create snapshots will have
surround commits that reference the snapshot branches, so the tool will
still create the appropriate git tag for the snapshots in the export
phase.