Fix branching off tag logic in export phase
If a branch's root is a tag then we have to use a tag dict to find the
mark number of the tag's commit to properly branch off the tag.

The logic for this was broken as the tag dict was set with the key of
the git safe name of the branch. Now we refer to branches in the dict
as the origional surround name of the branch.