Parse snapshot branches' files again
Snapshot branches are like tags, so it was thought that these branches
are always static in surround. Therefore there would be no reason to
parse these branches.

However, it appears in at least one project snapshot branches are
modified. Now we can again parse snapshot branches to detect these, but ignore
the add from branch operation if it is detected.

This behaviour only happens if --snapshot is thrown at the cli because it
should only be done for repos where it is necessary as it takes much longer
to parse with this behavior on.