Handle Surroudn branches that were renamed
Add to branch file operations always reference the old branch name if
a branch is renamed. This makes the script incorrectly view the branch
as deleted.

Now we also parse for branch renames in the parse phase, and fixup the
add to branch target branch to use the current name.