Refactor branch information into a branch class in parse phase
The main motivation behind this change is that we can place the
branch type into the new Branch class as soon as we discover all the
branches found. This allows us to remove the is_snaphshot() function usage in
the parse phase.

As an added benefit the arg list in the parse functions are cleaner now.