Fix finding branches in parse phase
The script previously assumed that the branch path and the repo path would
always be the same but this not actually true. However it is often true.

Now we fixup the logic to find main_branch and all its children so we always
find them correctly even when branch path != repo path

As an added benefit we dont need the path cli arg anymore. We can find the
branches with just the main branch and the mainline! So the cli args have
changed. That is probably a crazy thing to just throw into this commit, but
I am probably the only one using this tool anyway.