Use sscm property to properly find deleted files
This command output a list of deleted files and repos in a given repo.
Unfortunately this command has two problems that make it a bear to parse
this output.

1. Nothing in the output tells you if a deleted item is a file or repo

2. The deletion comment is also outputed on the same damn row as the
deleted item. It is impossible to make a regex that gets only the file
and not the comment.

To get around these we rerun the property command until it properly works.
This has a performance cost, but that is how it must be.