Update export phase to use a git fast-import subprocess
Instead of piping the output of this tool into git fast-import manually
use a subprocess to spawn git fast-import from the tool and send data
to this processes stdout.

The main benefit of this change is that we will be able to provide better
logs as we won't just have to print to stderr to avoid contaminating
gitfi.